Rename python scripts
