+++ update_client_configuration.py
